+++ CHANGES.rst
@@ -41,6 +41,7 @@ Changelog
 - added `count-specks` filter that adds counts of small objects to meta-data
 - added support for caching plugins via `IDC_CLASS_CACHE` environment variable
 - `od-to-is` now handles labels or regexp pattern correctly
+- added `is-to-od` filter that generates object detection annotations from contours determined in image segmentation layers